# C#: Refactor extractor trap generation code