--- a/csharp/extractor/Semmle.Extraction.CIL/Entities/Attribute.cs
+++ b/csharp/extractor/Semmle.Extraction.CIL/Entities/Attribute.cs
@@ -14,8 +14,9 @@ namespace Semmle.Extraction.CIL.Entities
     /// <summary>
     /// Entity representing a CIL attribute.
     /// </summary>
-    class Attribute : UnlabelledEntity, IAttribute
+    sealed class Attribute : UnlabelledEntity, IAttribute
     {
+        readonly CustomAttributeHandle handle;
         readonly CustomAttribute attrib;
         readonly IEntity @object;
 
@@ -24,6 +25,13 @@ namespace Semmle.Extraction.CIL.Entities
             attrib = cx.mdReader.GetCustomAttribute(handle);
             this.@object = @object;
         }
+
+        public override bool Equals(object obj)
+        {
+            return obj is Attribute attribute && handle.Equals(attribute.handle);
+        }
+
+        public override int GetHashCode() => handle.GetHashCode();
 
         public override IEnumerable<IExtractionProduct> Contents
         {
@@ -78,7 +86,7 @@ namespace Semmle.Extraction.CIL.Entities
         readonly Context cx;
         public CustomAttributeDecoder(Context cx) { this.cx = cx; }
 
-        public Type GetPrimitiveType(PrimitiveTypeCode typeCode) => cx.Populate(new PrimitiveType(cx, typeCode));
+        public Type GetPrimitiveType(PrimitiveTypeCode typeCode) => cx.Create(typeCode);
 
         public Type GetSystemType() => throw new NotImplementedException();