--- a/csharp/extractor/Semmle.Extraction.CIL/Entities/Event.cs
+++ b/csharp/extractor/Semmle.Extraction.CIL/Entities/Event.cs
@@ -7,7 +7,7 @@ namespace Semmle.Extraction.CIL.Entities
     /// <summary>
     /// An event.
     /// </summary>
-    interface IEvent : ILabelledEntity
+    interface IEvent : IExtractedEntity
     {
     }