--- a/csharp/extractor/Semmle.Extraction.CIL/Entities/Field.cs
+++ b/csharp/extractor/Semmle.Extraction.CIL/Entities/Field.cs
@@ -13,7 +13,7 @@ namespace Semmle.Extraction.CIL.Entities
     /// An entity represting a member.
     /// Used to type tuples correctly.
     /// </summary>
-    interface IMember : ILabelledEntity
+    interface IMember : IExtractedEntity
     {
     }