--- a/csharp/extractor/Semmle.Extraction.CIL/Entities/Folder.cs
+++ b/csharp/extractor/Semmle.Extraction.CIL/Entities/Folder.cs
@@ -16,7 +16,6 @@ namespace Semmle.Extraction.CIL.Entities
             this.path = path;
         }
 
-
         public override void WriteId(TextWriter trapFile)
         {
             trapFile.Write(Semmle.Extraction.Entities.File.PathAsDatabaseId(path));