--- a/csharp/extractor/Semmle.Extraction.CIL/Entities/LocalVariable.cs
+++ b/csharp/extractor/Semmle.Extraction.CIL/Entities/LocalVariable.cs
@@ -3,7 +3,7 @@ using System.IO;
 
 namespace Semmle.Extraction.CIL.Entities
 {
-    interface ILocal : ILabelledEntity
+    interface ILocal : IExtractedEntity
     {
     }