--- a/csharp/extractor/Semmle.Extraction.CIL/Entities/Parameter.cs
+++ b/csharp/extractor/Semmle.Extraction.CIL/Entities/Parameter.cs
@@ -6,7 +6,7 @@ namespace Semmle.Extraction.CIL.Entities
     /// <summary>
     /// A parameter entity.
     /// </summary>
-    interface IParameter : ILabelledEntity
+    interface IParameter : IExtractedEntity
     {
     }