- Option for Brotli compression
- Refactor Id generation in C# and CIL extractors, using `WriteId` to write directly to the trap file
- Refactor tuple generation in C# to write directly to the trap file
- Various code tidying.

Improves the clarity and performance of the code for generating trap output. The main strategy has been to write to `TextWriter` directly instead of adding intermediate abstractions that did not do very much to improve maintainability and seemed to hurt performance.  For example

```
cx.Emit(Tuples.methods(....))
```
has been rewritten to
```
trapFile.methods(....)
```
where the `methods` extension method writes the tuple directly to the `TextWriter` instead of constructing an intermediate object that then wrote itself to an `ITrapWriter`.

IDs are similarly simplified, using a `WriteId(TextWriter trapFile)` method throughout, allowing objects to write their IDs to trap files directly instead of via intermediate temporary objects. The classes in the CIL extractor relating to ID construction have been removed in favour of the new approach, making the two extractors much more similar.

The resulting code is about 15% faster in extraction time (total time spent in the extractor) in both the single threaded and multi-threaded case.

C#+CIL speedups:
-  Powershell: 112s->99s (-13%)
-  C# extractor: 63s -> 52s (-21%)

C# only speedups: 
-  Poweshell: 127s -> 109s (-16%)
-  C# extractor: 41s -> 38s (-7%)

Brotli compression can be configured by passing `--brotli` to the extractor, but it is disabled by default because it does not give a speed increase, and `.br` files are harder to work with. It does however give a reduction of trap size of around 20% (95MB -> 78MB)