@hvitved 2019-08-29
Thanks for addressing many of my comments, I have marked the relevant comments as resolved.